refactor(library): move reserve_copy method and can_make_reservation methods from models.py to services.py
